Implemented screen which show the solutions, modified certainty factor threshold value from 20 to 40, implemented reset functionality
Implemented screen which shows the solutions after the consulting
process. If there are no solutions, a message will be displayed. The
solutions will be displayed in a list, in descending order of the
certainty factor. Also, I created the Solution class, which models a solution.

After viewing the solutions, the user can click Reset system button,
which will bring him to the main menu, and he will be able to consult
the system again.

I also changed the certainty factor threshold from 20 to 40.